third_party/blink/web_tests/platform/generic/fullscreen/full-screen-cancel-nested-expected.txt

Test that webkitCancelFullScreen only exits one level of fullscreen

EVENT(webkitfullscreenchange)
EXPECTED (document.webkitCurrentFullScreenElement == '[object HTMLDivElement]') OK
EVENT(webkitfullscreenchange)
EXPECTED (document.webkitCurrentFullScreenElement == '[object HTMLSpanElement]') OK
EVENT(webkitfullscreenchange)
EXPECTED (document.webkitCurrentFullScreenElement == '[object HTMLDivElement]') OK
END OF TEST
